reorganize e2e tests
